backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.11
python-decouple==3.8
drf-spectacular==0.27.1
openpyxl==3.1.2
reportlab==4.0.9
Pillow==11.2.1
python-dateutil==2.9.0
gunicorn==22.0.0
